Refactored nn.basis to nn.weight_functions, with basis weight functions as one option (only one implemented so far).
This also changes local and gradient to be their own weighted density, rather than being mixed in as in previous case.

+++ src/mldft1d/nn/function.py
@@ -81,7 +81,7 @@ class Linear(torch.nn.Module):
             torch.nn.init.uniform_(self.bias, -bound, bound)
 
     def forward(self, x: torch.Tensor) -> torch.Tensor:
-        result = torch.tensordot(self.weight.to(x.dtype), x, dims=1)
+        result = torch.tensordot(self.weight, x, dims=1)
         if self.bias is not None:
             result += self.bias.view((self.n_out,) + (1,) * (len(x.shape) - 1))
         return result
